# core/openapi.yaml
openapi: 3.0.0
info:
  title: WACAST Core API
  description: WhatsApp Multi-Session Gateway API - Enterprise WhatsApp messaging platform
  version: 1.0.0
  contact:
    name: WACAST Support
    email: support@wacast.io
  license:
    name: Apache 2.0
servers:
  - url: http://localhost:8080
    description: Local Development Server
  - url: https://api.wacast.io
    description: Production Server
...
paths:
...
  /api/v1/devices/{device_id}/messages:
    post:
      tags:
        - Messages
      summary: Send text message
      description: Send a text message via an active WhatsApp session
      operationId: sendMessage
      parameters:
        - name: device_id
          in: path
          required: true
          schema:
            type: string
          description: Device/Session identifier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageRequest"
            example:
              target_jid: "62812345678"
              content: "Hello, this is a test message"
              group_id: null
      responses:
        "202":
          description: Message accepted and queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid request body
        "404":
          description: Session not found
  /api/v1/devices/{device_id}/messages/media:
    post:
      tags:
        - Messages
      summary: Send media message
      description: Send an image, document, audio, or video message
      operationId: sendMediaMessage
      parameters:
        - name: device_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMediaMessageRequest"
            example:
              target_jid: "62812345678"
              media_url: "https://example.com/image.jpg"
              content_type: "image"
              caption: "Check this out!"
      responses:
        "202":
          description: Message accepted and queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /api/v1/devices/{device_id}/messages/scheduled:
    post:
      tags:
        - Messages
      summary: Schedule message
      description: Schedule a message to be sent at a specific time
      operationId: sendScheduledMessage
      parameters:
        - name: device_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendScheduledMessageRequest"
            example:
              target_jid: "62812345678"
              content: "Good morning!"
              scheduled_for: "2026-04-10T08:00:00Z"
      responses:
        "202":
          description: Message scheduled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /api/v1/messages/{message_id}/status:
    get:
      tags:
        - Messages
      summary: Get message status
      description: Check delivery and read status of a sent message
      operationId: getMessageStatus
      parameters:
        - name: message_id
          in: path
          required: true
          schema:
            type: string
          description: Message identifier
      responses:
        "200":
          description: Message status retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageStatusResponse"
        "404":
          description: Message not found
...
  /api/v1/messages/failed:
    get:
      tags:
        - Messages
      summary: List failed messages
      description: Returns list of messages that failed to deliver
      operationId: getFailedMessages
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
          description: Maximum number of results
      responses:
        "200":
          description: List of failed messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FailedMessagesResponse"
...
components:
  schemas:
...
    SendMessageRequest:
      type: object
      required:
        - target_jid
        - content
      properties:
        target_jid:
          type: string
          description: Recipient phone number or JID
        content:
          type: string
          description: Message content
        group_id:
          type: string
          nullable: true
          description: Group ID if sending to group
    SendMediaMessageRequest:
      type: object
      required:
        - target_jid
        - media_url
        - content_type
      properties:
        target_jid:
          type: string
        media_url:
          type: string
          description: URL to media file
        content_type:
          type: string
          enum: [image, document, audio, video, contact, location]
        caption:
          type: string
          nullable: true
          description: Optional caption for media
    SendScheduledMessageRequest:
      type: object
      required:
        - target_jid
        - content
        - scheduled_for
      properties:
        target_jid:
          type: string
        content:
          type: string
        scheduled_for:
          type: string
          format: date-time
          description: ISO 8601 datetime when to send
    MessageResponse:
      type: object
      properties:
        message_id:
          type: string
        status:
          type: string
          enum: [pending, sent, delivered, read, failed]
        timestamp:
          type: integer
    MessageStatusResponse:
      type: object
      properties:
        message_id:
          type: string
        status:
          type: string
          enum: [pending, sent, delivered, read, failed]
        timestamp:
          type: integer
...
    FailedMessagesResponse:
      type: object
      properties:
        count:
          type: integer
        messages:
          type: array
          items:
            type: object
            properties:
              message_id:
                type: string
              device_id:
                type: string
              target_jid:
                type: string
              content:
                type: string
              retry_count:
                type: integer
              max_retries:
                type: integer
              error:
                type: string
...